moving the function in utils so that can be used by scanOld script
